More changes from vkQuake2 by @kondrak.

+++ m_parasite.c
@@ -336,7 +336,7 @@ void parasite_drain_attack (edict_t *self)
 
 	gi.WriteByte (svc_temp_entity);
 	gi.WriteByte (TE_PARASITE_ATTACK);
-	gi.WriteShort ((int)(self - g_edicts));
+	gi.WriteShort (self - g_edicts);
 	gi.WritePosition (start);
 	gi.WritePosition (end);
 	gi.multicast (self->s.origin, MULTICAST_PVS);
@@ -513,7 +513,7 @@ qboolean parasite_blocked (edict_t *self, float dist)
 
 	if(blocked_checkplat (self, dist))
 		return true;
-		
+
 	return false;
 }
 //ROGUE
@@ -559,7 +559,7 @@ qboolean parasite_checkattack (edict_t *self)
 		self->monsterinfo.aiflags &= ~AI_BLOCKED;
 		return true;
 	}
-	
+
 	return false;
 }
