+++ p_client.c
@@ -1088,9 +1088,9 @@ edict_t *SelectLavaCoopSpawnPoint (edict_t *ent)
 
 edict_t *SelectCoopSpawnPoint (edict_t *ent)
 {
-	intptr_t	index;
-	edict_t		*spot = NULL;
-	char		*target;
+	int		index;
+	edict_t	*spot = NULL;
+	char	*target;
 
 //ROGUE
 	// rogue hack, but not too gross...
@@ -1221,7 +1221,7 @@ void CopyToBodyQue (edict_t *ent)
 
 	gi.unlinkentity (body);
 	body->s = ent->s;
-	body->s.number = (int)(body - g_edicts);
+	body->s.number = body - g_edicts;
 
 	body->svflags = ent->svflags;
 	VectorCopy (ent->mins, body->mins);
@@ -1339,7 +1339,7 @@ void spectator_respawn (edict_t *ent)
 	{
 		// send effect
 		gi.WriteByte (svc_muzzleflash);
-		gi.WriteShort ((int)(ent-g_edicts));
+		gi.WriteShort (ent-g_edicts);
 		gi.WriteByte (MZ_LOGIN);
 		gi.multicast (ent->s.origin, MULTICAST_PVS);
 
@@ -1369,12 +1369,12 @@ a deathmatch.
 */
 void PutClientInServer (edict_t *ent)
 {
-	vec3_t		mins = {-16, -16, -24};
-	vec3_t		maxs = {16, 16, 32};
-	intptr_t	index;
-	vec3_t		spawn_origin, spawn_angles;
+	vec3_t	mins = {-16, -16, -24};
+	vec3_t	maxs = {16, 16, 32};
+	int		index;
+	vec3_t	spawn_origin, spawn_angles;
 	gclient_t	*client;
-	int			i;
+	int		i;
 	client_persistant_t	saved;
 	client_respawn_t	resp;
 
@@ -1495,7 +1495,7 @@ void PutClientInServer (edict_t *ent)
 	ent->s.modelindex2 = 255;		// custom gun model
 	// sknum is player num and weapon number
 	// weapon number will be added in changeweapon
-	ent->s.skinnum = (int)(ent - g_edicts - 1);
+	ent->s.skinnum = ent - g_edicts - 1;
 
 	ent->s.frame = 0;
 	VectorCopy (spawn_origin, ent->s.origin);
@@ -1588,7 +1588,7 @@ void ClientBeginDeathmatch (edict_t *ent)
 	{
 		// send effect
 		gi.WriteByte (svc_muzzleflash);
-		gi.WriteShort ((int)(ent-g_edicts));
+		gi.WriteShort (ent-g_edicts);
 		gi.WriteByte (MZ_LOGIN);
 		gi.multicast (ent->s.origin, MULTICAST_PVS);
 	}
@@ -1652,7 +1652,7 @@ void ClientBegin (edict_t *ent)
 		if (game.maxclients > 1)
 		{
 			gi.WriteByte (svc_muzzleflash);
-			gi.WriteShort ((int)(ent-g_edicts));
+			gi.WriteShort (ent-g_edicts);
 			gi.WriteByte (MZ_LOGIN);
 			gi.multicast (ent->s.origin, MULTICAST_PVS);
 
@@ -1701,7 +1701,7 @@ void ClientUserinfoChanged (edict_t *ent, char *userinfo)
 	// set skin
 	s = Info_ValueForKey (userinfo, "skin");
 
-	playernum = (int)(ent-g_edicts-1);
+	playernum = ent-g_edicts-1;
 
 	// combine name and skin into a configstring
 	gi.configstring (CS_PLAYERSKINS+playernum, va("%s\\%s", ent->client->pers.netname, s) );
@@ -1860,7 +1860,7 @@ void ClientDisconnect (edict_t *ent)
 
 	// send effect
 	gi.WriteByte (svc_muzzleflash);
-	gi.WriteShort ((int)(ent-g_edicts));
+	gi.WriteShort (ent-g_edicts);
 	gi.WriteByte (MZ_LOGOUT);
 	gi.multicast (ent->s.origin, MULTICAST_PVS);
 
@@ -1871,7 +1871,7 @@ void ClientDisconnect (edict_t *ent)
 	ent->classname = "disconnected";
 	ent->client->pers.connected = false;
 
-	playernum = (int)(ent-g_edicts-1);
+	playernum = ent-g_edicts-1;
 	gi.configstring (CS_PLAYERSKINS+playernum, "");
 }
